Added more data files to JSONReplies
Added a RequestUrlList which contains every possible URL that can be sent to the Battle.net API as of today.

Also added a WoWEURealmsList, and a WoWAuctionData.json. The latter contains a response from the Alteral Mountains auction data API, it's got every auction running at the time I executed the query.