Beginning to add test classes to verify integrity
To help verify integrity of the library and that it's generating the correct URL's, I added a test class that tests the WoWRequest class to ensure it's generating valid URL's.

Next is to add a test class for AlgalonClient that verifies the generated URL's actually return data in case Blizzard modifies the API in some way.